moved explode() definition from GeoPandasBase to GeoSeries
