+++ geopandas/base.py
@@ -1200,53 +1200,53 @@ GeometryCollection
             "skew", self, xs, ys, origin=origin, use_radians=use_radians
         )
 
-    def explode(self):
-        """
-        Explode multi-part geometries into multiple single geometries.
-
-        Single rows can become multiple rows.
-        This is analogous to PostGIS's ST_Dump(). The 'path' index is the
-        second level of the returned MultiIndex
-
-        Returns
-        ------
-        A GeoSeries with a MultiIndex. The levels of the MultiIndex are the
-        original index and a zero-based integer index that counts the
-        number of single geometries within a multi-part geometry.
-
-        Examples
-        --------
-        >>> from shapely.geometry import MultiPoint
-        >>> s = geopandas.GeoSeries(
-        ...     [MultiPoint([(0, 0), (1, 1)]), MultiPoint([(2, 2), (3, 3), (4, 4)])]
-        ... )
-        >>> s
-        0        MULTIPOINT (0.00000 0.00000, 1.00000 1.00000)
-        1    MULTIPOINT (2.00000 2.00000, 3.00000 3.00000, ...
-        dtype: geometry
-
-        >>> s.explode()
-        0  0    POINT (0.00000 0.00000)
-           1    POINT (1.00000 1.00000)
-        1  0    POINT (2.00000 2.00000)
-           1    POINT (3.00000 3.00000)
-           2    POINT (4.00000 4.00000)
-        dtype: geometry
-
-        """
-        index = []
-        geometries = []
-        for idx, s in self.geometry.iteritems():
-            if s.type.startswith("Multi") or s.type == "GeometryCollection":
-                geoms = s.geoms
-                idxs = [(idx, i) for i in range(len(geoms))]
-            else:
-                geoms = [s]
-                idxs = [(idx, 0)]
-            index.extend(idxs)
-            geometries.extend(geoms)
-        index = MultiIndex.from_tuples(index, names=self.index.names + [None])
-        return gpd.GeoSeries(geometries, index=index, crs=self.crs).__finalize__(self)
+    # def explode(self):
+    #     """
+    #     Explode multi-part geometries into multiple single geometries.
+    #
+    #     Single rows can become multiple rows.
+    #     This is analogous to PostGIS's ST_Dump(). The 'path' index is the
+    #     second level of the returned MultiIndex
+    #
+    #     Returns
+    #     ------
+    #     A GeoSeries with a MultiIndex. The levels of the MultiIndex are the
+    #     original index and a zero-based integer index that counts the
+    #     number of single geometries within a multi-part geometry.
+    #
+    #     Examples
+    #     --------
+    #     >>> from shapely.geometry import MultiPoint
+    #     >>> s = geopandas.GeoSeries(
+    #     ...     [MultiPoint([(0, 0), (1, 1)]), MultiPoint([(2, 2), (3, 3), (4, 4)])]
+    #     ... )
+    #     >>> s
+    #     0        MULTIPOINT (0.00000 0.00000, 1.00000 1.00000)
+    #     1    MULTIPOINT (2.00000 2.00000, 3.00000 3.00000, ...
+    #     dtype: geometry
+    #
+    #     >>> s.explode()
+    #     0  0    POINT (0.00000 0.00000)
+    #        1    POINT (1.00000 1.00000)
+    #     1  0    POINT (2.00000 2.00000)
+    #        1    POINT (3.00000 3.00000)
+    #        2    POINT (4.00000 4.00000)
+    #     dtype: geometry
+    #
+    #     """
+    #     index = []
+    #     geometries = []
+    #     for idx, s in self.geometry.iteritems():
+    #         if s.type.startswith("Multi") or s.type == "GeometryCollection":
+    #             geoms = s.geoms
+    #             idxs = [(idx, i) for i in range(len(geoms))]
+    #         else:
+    #             geoms = [s]
+    #             idxs = [(idx, 0)]
+    #         index.extend(idxs)
+    #         geometries.extend(geoms)
+    #     index = MultiIndex.from_tuples(index, names=self.index.names + [None])
+    #     return gpd.GeoSeries(geometries, index=index, crs=self.crs).__finalize__(self)
 
     @property
     def cx(self):
